Cleanup of examples.
